requirements.txt
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
et-xmlfile==1.1.0
google-analytics-data==0.16.0
google-api-core==2.11.0
google-auth==2.16.0
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.2
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
rsa==4.9
six==1.16.0
urllib3==1.26.14
